finished refactoring create.js
