+++ browser/game/engine/create.js
@@ -3,9 +3,9 @@ import Survivor from '../controls/survivor.js';
 import GameMaster from '../controls/gameMaster.js';
 import socket from '../../socket'
 import Bullets from '../entities/bullets.js';
+import { createMapPrePlayer, createMapPostPlayer } from './createMap.js'
 
-let player, bullets, playerCollide, teamBullet, survivor, gameMaster;
-let teammates = {}
+let player, bullets, teamBullet, survivor, gameMaster;
 
 export default function create() {
   //this settings
@@ -13,20 +13,7 @@ export default function create() {
   this.physics.startSystem(Phaser.Physics.ARCADE);
 
   // map, order matters!
-  let map = this.add.tilemap('tilemap');
-  map.addTilesetImage('terrain_atlas', 'tileset');
-
-  playerCollide = map.createLayer('playerCollide');
-
-  let groundLayer = map.createLayer('groundLayer');
-  let featuresBottom = map.createLayer('featuresBottom');
-
-  groundLayer.resizeWorld();
-  map.setCollisionBetween(1, 2000, true, 'playerCollide');
-
-  let playerOnBottom = map.createLayer('playerOnBottom');
-  let playerOnTop = map.createLayer('playerOnTop');
-  //     player = new Player(socket.id, this);
+  createMapPrePlayer(this)
 
   if (store.getState().gameMode === 'survivor') {
     survivor = new Survivor(this);
@@ -36,12 +23,9 @@ export default function create() {
     gameMaster = new GameMaster(this);
   }
 
-  let playerBehindBottom = map.createLayer('playerBehindBottom');
-  let playerBehindTop = map.createLayer('playerBehindTop');
+  createMapPostPlayer()
 
   teamBullet = new Bullets(this);
-  //button
-  //button = this.add.button(this.world.centerX - 95, 400, 'button', spawn, this, 2, 1, 0);
 
   const emitClient = () => {
     setInterval(() => {
@@ -52,4 +36,4 @@ export default function create() {
   emitClient()
 }
 
-export {player, bullets, teammates, teamBullet, playerCollide, survivor, gameMaster};
+export {player, bullets, teamBullet, survivor, gameMaster};
