+++ browser/game/engine/update.js
@@ -1,9 +1,12 @@
-import {bullets, teammates, player, playerCollide} from './create.js';
+import { bullets, player } from './create.js';
+import { playerCollide } from './createMap.js';
 import { monsters, monstersLocation } from '../controls/controls.js';
 import Teammate from '../entities/teammate.js';
 import store from '../../store.js';
 import { receiveHealth } from '../../reducers/players.js';
 import GameMaster from '../controls/gameMaster.js';
+
+let teammates = {};
 
 export default function update() {
   //  Collision
